[power_switching_tools_ros] Make it easier to support new device
