+++ power_switching_tools_ros/setup.py
@@ -1,0 +1,11 @@
+from catkin_pkg.python_setup import generate_distutils_setup
+from setuptools import find_packages
+from setuptools import setup
+
+
+d = generate_distutils_setup(
+    packages=find_packages('src'),
+    package_dir={'': 'src'},
+)
+
+setup(**d)
